+++ internal/gengapic/genrest.go
@@ -162,6 +162,15 @@ func (g *generator) restClientUtilities(serv *descriptorpb.ServiceDescriptorProt
 	g.serviceDoc(serv, false) // exclude API version docs
 	p("func New%[1]sRESTClient(ctx context.Context, opts ...option.ClientOption) (*%[1]sClient, error) {", servName)
 	p("    clientOpts := append(default%sRESTClientOptions(), opts...)", servName)
+	if g.featureEnabled(OpenTelemetryTracingFeature) {
+		p("    if gax.IsFeatureEnabled(\"TRACING\") {")
+		p("        clientOpts = append(clientOpts, internaloption.WithTelemetryAttributes(map[string]string{")
+		p("          \"gcp.client.service\": %q,", strings.Split(g.cfg.APIServiceConfig.GetName(), ".")[0])
+		p("          \"gcp.client.version\": getVersionClient(),")
+		p("          \"gcp.client.repo\":    \"googleapis/google-cloud-go\",")
+		p("        }))")
+		p("    }")
+	}
 	p("    httpClient, endpoint, err := httptransport.NewClient(ctx, clientOpts...)")
 	p("    if err != nil {")
 	p("        return nil, err")
@@ -657,6 +666,12 @@ func (g *generator) serverStreamRESTCall(servName string, s *descriptorpb.Servic
 	g.generateQueryString(m)
 	p("// Build HTTP headers from client and context metadata.")
 	g.insertRequestHeaders(m, rest)
+	if info != nil && g.featureEnabled(OpenTelemetryTracingFeature) {
+		p("if gax.IsFeatureEnabled(\"TRACING\") {")
+		p("  ctx = metadata.AppendToOutgoingContext(ctx, \"url.template\", %q)", info.url)
+		p("}")
+		g.imports[pbinfo.ImportSpec{Path: "google.golang.org/grpc/metadata"}] = true
+	}
 	p("var streamClient *%s", streamClient)
 	p("e := gax.Invoke(ctx, func(ctx context.Context, settings gax.CallSettings) error {")
 	p(`  if settings.Path != "" {`)
@@ -939,6 +954,12 @@ func (g *generator) lroRESTCall(servName string, m *descriptorpb.MethodDescripto
 	g.generateQueryString(m)
 	p("// Build HTTP headers from client and context metadata.")
 	g.insertRequestHeaders(m, rest)
+	if info != nil && g.featureEnabled(OpenTelemetryTracingFeature) {
+		p("if gax.IsFeatureEnabled(\"TRACING\") {")
+		p("  ctx = metadata.AppendToOutgoingContext(ctx, \"url.template\", %q)", info.url)
+		p("}")
+		g.imports[pbinfo.ImportSpec{Path: "google.golang.org/grpc/metadata"}] = true
+	}
 	p("unm := protojson.UnmarshalOptions{AllowPartial: true, DiscardUnknown: true}")
 	p("resp := &%s.%s{}", outSpec.Name, outType.GetName())
 	p("e := gax.Invoke(ctx, func(ctx context.Context, settings gax.CallSettings) error {")
@@ -1035,6 +1056,12 @@ func (g *generator) emptyUnaryRESTCall(servName string, m *descriptorpb.MethodDe
 	g.generateQueryString(m)
 	p("// Build HTTP headers from client and context metadata.")
 	g.insertRequestHeaders(m, rest)
+	if info != nil && g.featureEnabled(OpenTelemetryTracingFeature) {
+		p("if gax.IsFeatureEnabled(\"TRACING\") {")
+		p("  ctx = metadata.AppendToOutgoingContext(ctx, \"url.template\", %q)", info.url)
+		p("}")
+		g.imports[pbinfo.ImportSpec{Path: "google.golang.org/grpc/metadata"}] = true
+	}
 	p("return gax.Invoke(ctx, func(ctx context.Context, settings gax.CallSettings) error {")
 	p(`  if settings.Path != "" {`)
 	p("    baseUrl.Path = settings.Path")
@@ -1124,6 +1151,12 @@ func (g *generator) unaryRESTCall(servName string, m *descriptorpb.MethodDescrip
 	g.generateQueryString(m)
 	p("// Build HTTP headers from client and context metadata.")
 	g.insertRequestHeaders(m, rest)
+	if info != nil && g.featureEnabled(OpenTelemetryTracingFeature) {
+		p("if gax.IsFeatureEnabled(\"TRACING\") {")
+		p("  ctx = metadata.AppendToOutgoingContext(ctx, \"url.template\", %q)", info.url)
+		p("}")
+		g.imports[pbinfo.ImportSpec{Path: "google.golang.org/grpc/metadata"}] = true
+	}
 	g.appendCallOpts(m)
 	if !isHTTPBodyMessage {
 		p("unm := protojson.UnmarshalOptions{AllowPartial: true, DiscardUnknown: true}")