+++ internal/gengapic/genrest_test.go
@@ -440,10 +440,15 @@ func TestGenRestMethod(t *testing.T) {
 		Type:    descriptorpb.FieldDescriptorProto_TYPE_INT32.Enum(),
 		Options: sizeOpts,
 	}
+	resRefOpts := &descriptorpb.FieldOptions{}
+	proto.SetExtension(resRefOpts, annotations.E_ResourceReference, &annotations.ResourceReference{
+		Type: "foo.googleapis.com/Bar",
+	})
 	otherField := &descriptorpb.FieldDescriptorProto{
 		Name:           proto.String("other"),
 		Type:           descriptorpb.FieldDescriptorProto_TYPE_STRING.Enum(),
 		Proto3Optional: proto.Bool(true),
+		Options:        resRefOpts,
 	}
 	infoOpts := &descriptorpb.FieldOptions{}
 	proto.SetExtension(infoOpts, annotations.E_FieldInfo, &annotations.FieldInfo{Format: annotations.FieldInfo_UUID4})
@@ -673,8 +678,10 @@ func TestGenRestMethod(t *testing.T) {
 	}
 
 	s := &descriptorpb.ServiceDescriptorProto{
-		Name: proto.String("FooService"),
+		Name:    proto.String("FooService"),
+		Options: &descriptorpb.ServiceOptions{},
 	}
+	proto.SetExtension(s.Options, annotations.E_DefaultHost, "foo.googleapis.com")
 	opS := &descriptorpb.ServiceDescriptorProto{
 		Name: proto.String("FooOperationService"),
 	}
@@ -696,7 +703,7 @@ func TestGenRestMethod(t *testing.T) {
 			methodToWrapper: map[*descriptorpb.MethodDescriptorProto]operationWrapper{},
 			opWrappers:      map[string]operationWrapper{},
 		},
-		cfg: &generatorConfig{},
+		cfg: &generatorConfig{featureEnablement: map[featureID]struct{}{OpenTelemetryTracingFeature: {}}},
 		customOpServices: map[*descriptorpb.ServiceDescriptorProto]*descriptorpb.ServiceDescriptorProto{
 			s: opS,
 		},
@@ -753,8 +760,9 @@ func TestGenRestMethod(t *testing.T) {
 		{
 			name:   "custom_op",
 			method: opRPC,
-			cfg:    &generatorConfig{generateAsDIREGAPIC: true},
+			cfg:    &generatorConfig{generateAsDIREGAPIC: true, featureEnablement: map[featureID]struct{}{OpenTelemetryTracingFeature: {}}},
 			imports: map[pbinfo.ImportSpec]bool{
+				{Path: "google.golang.org/grpc/metadata"}:               true,
 				{Path: "google.golang.org/protobuf/encoding/protojson"}: true,
 				{Path: "net/url"}: true,
 				{Path: "fmt"}:     true,
@@ -764,33 +772,35 @@ func TestGenRestMethod(t *testing.T) {
 		{
 			name:   "empty_rpc",
 			method: emptyRPC,
-			cfg:    &generatorConfig{},
+			cfg:    &generatorConfig{featureEnablement: map[featureID]struct{}{OpenTelemetryTracingFeature: {}}},
 			imports: map[pbinfo.ImportSpec]bool{
-				{Path: "fmt"}:                    true,
-				{Path: "github.com/google/uuid"}: true,
-				{Path: "net/url"}:                true,
+				{Path: "fmt"}:                             true,
+				{Path: "github.com/google/uuid"}:          true,
+				{Path: "google.golang.org/grpc/metadata"}: true,
+				{Path: "net/url"}:                         true,
 				{Name: "foopb", Path: "google.golang.org/genproto/cloud/foo/v1"}: true,
 			},
 		},
 		{
 			name:   "unary_rpc",
 			method: unaryRPC,
-			cfg:    &generatorConfig{restNumericEnum: true},
+			cfg:    &generatorConfig{restNumericEnum: true, featureEnablement: map[featureID]struct{}{OpenTelemetryTracingFeature: {}}},
 			imports: map[pbinfo.ImportSpec]bool{
-				{Path: "bytes"}:                  true,
-				{Path: "fmt"}:                    true,
-				{Path: "github.com/google/uuid"}: true,
+				{Path: "bytes"}:                                         true,
+				{Path: "fmt"}:                                           true,
+				{Path: "github.com/google/uuid"}:                        true,
+				{Path: "google.golang.org/grpc/metadata"}:               true,
 				{Path: "google.golang.org/protobuf/encoding/protojson"}: true,
-				{Path: "net/url"}: true,
-				{Path: "regexp"}:  true,
-				{Path: "strings"}: true,
+				{Path: "net/url"}:                                       true,
+				{Path: "regexp"}:                                        true,
+				{Path: "strings"}:                                       true,
 				{Name: "foopb", Path: "google.golang.org/genproto/cloud/foo/v1"}: true,
 			},
 		},
 		{
 			name:   "paging_rpc",
 			method: pagingRPC,
-			cfg:    &generatorConfig{},
+			cfg:    &generatorConfig{featureEnablement: map[featureID]struct{}{OpenTelemetryTracingFeature: {}}},
 			imports: map[pbinfo.ImportSpec]bool{
 				{Path: "math"}:    true,
 				{Path: "net/url"}: true,
@@ -804,7 +814,7 @@ func TestGenRestMethod(t *testing.T) {
 		{
 			name:   "server_stream_rpc",
 			method: serverStreamRPC,
-			cfg:    &generatorConfig{},
+			cfg:    &generatorConfig{featureEnablement: map[featureID]struct{}{OpenTelemetryTracingFeature: {}}},
 			imports: map[pbinfo.ImportSpec]bool{
 				{Path: "bytes"}:   true,
 				{Path: "context"}: true,
@@ -821,7 +831,7 @@ func TestGenRestMethod(t *testing.T) {
 		{
 			name:   "no_request_stream_rpc",
 			method: clientStreamRPC,
-			cfg:    &generatorConfig{},
+			cfg:    &generatorConfig{featureEnablement: map[featureID]struct{}{OpenTelemetryTracingFeature: {}}},
 			imports: map[pbinfo.ImportSpec]bool{
 				{Path: "context"}: true,
 				{Path: "errors"}:  true,
@@ -831,8 +841,9 @@ func TestGenRestMethod(t *testing.T) {
 		{
 			name:   "lro_rpc",
 			method: lroRPC,
-			cfg:    &generatorConfig{transports: []transport{rest}},
+			cfg:    &generatorConfig{transports: []transport{rest}, featureEnablement: map[featureID]struct{}{OpenTelemetryTracingFeature: {}}},
 			imports: map[pbinfo.ImportSpec]bool{
+				{Path: "google.golang.org/grpc/metadata"}: true,
 				{Path: "bytes"}: true,
 				{Path: "cloud.google.com/go/longrunning"}: true,
 				{Path: "fmt"}:                    true,
@@ -845,10 +856,11 @@ func TestGenRestMethod(t *testing.T) {
 		{
 			name:   "http_body_rpc",
 			method: httpBodyRPC,
-			cfg:    &generatorConfig{},
+			cfg:    &generatorConfig{featureEnablement: map[featureID]struct{}{OpenTelemetryTracingFeature: {}}},
 			imports: map[pbinfo.ImportSpec]bool{
 				{Path: "bytes"}: true,
 				{Path: "fmt"}:   true,
+				{Path: "google.golang.org/grpc/metadata"}:               true,
 				{Path: "google.golang.org/protobuf/encoding/protojson"}: true,
 				{Path: "net/url"}: true,
 				{Path: "regexp"}:  true,
@@ -860,8 +872,9 @@ func TestGenRestMethod(t *testing.T) {
 		{
 			name:   "update_rpc",
 			method: updateRPC,
-			cfg:    &generatorConfig{restNumericEnum: true},
+			cfg:    &generatorConfig{restNumericEnum: true, featureEnablement: map[featureID]struct{}{OpenTelemetryTracingFeature: {}}},
 			imports: map[pbinfo.ImportSpec]bool{
+				{Path: "google.golang.org/grpc/metadata"}: true,
 				{Path: "bytes"}: true,
 				{Path: "fmt"}:   true,
 				{Path: "google.golang.org/protobuf/encoding/protojson"}: true,